scraping/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
banks==2.2.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
filelock==3.19.1
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html_text==0.7.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.2
llama-index-cli==0.5.1
llama-index-core==0.14.2
llama-index-embeddings-google-genai==0.3.1
llama-index-embeddings-ollama==0.8.3
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-google-genai==0.5.1
llama-index-llms-ollama==0.7.3
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-qdrant==0.8.5
llama-index-vector-stores-weaviate==1.4.1
llama-index-workflows==2.2.2
llama-parse==0.6.54
looseversion==1.3.0
lxml==6.0.2
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.3
ollama==0.5.4
openai==1.108.2
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.5
pypdf==6.1.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.15.1
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
weaviate-client==4.16.10
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
